ampps-tkb

+++ news/delete.php
@@ -3,11 +3,23 @@
 		if ($_POST['action'] == "delete") {
 			include '../php/init.php';
 			foreach ($_POST['id'] as $key => $id) {
-				$sql = "SELECT id FROM $tab WHERE id = '$id'";
+				$sql = "SELECT * FROM $tab WHERE id = '$id'";
 				$data = $db->query($sql);
 				if ($data->num_rows > 0) {
-					$sql = "DELETE FROM $tab WHERE id = '$id'";
-					$db->query($sql);
+					$row = $data->fetch_assoc();
+					$valid = False;
+					if ($tab == 'photos') {
+						if (unlink('../../'.$row['url'])) {
+							$valid = True;
+						}
+					}
+					else {
+						$valid = True;
+					}
+					if ($valid) {
+						$sql = "DELETE FROM $tab WHERE id = '$id'";
+						$db->query($sql);
+					}
 				}
 			}
 		}
